# Conflicts:
#	DDI/DDI.UI.Web/DDI.UI.Web.csproj

+++ DDI/DDI.UI.Web/Scripts/data.js
@@ -80,6 +80,10 @@ function PopulateDropDown(element, route, selectedValue) {
 
 function PopulateDropDown(element, route, defaultText, defaultValue, selectedValue, changecallback, completecallback) {
 
+    if ($.type(element) === "string" && element.indexOf('.') != 0) {
+        element = '.' + element;
+    }
+
     ClearElement(element);
 
     AddDefaultOption(element, defaultText, defaultValue);
@@ -186,44 +190,47 @@ function LoadGrid(container, gridClass, columns, getRoute, saveRoute, selected, 
         LoadGridData(container, gridClass, columns, getRoute, selected, newlink, showFilter, showGroup, onComplete);
     }
 
-    if (editModalClass) {
-        // Add column for edit
-        columns.push({
-            width: '100px',
-            alignment: 'center',
-            cellTemplate: function(container, options) {
-                $('<a/>')
-                    .addClass('editLink')
-                    .text('Edit')
-                    .click(function(e) {
-                        e.preventDefault();
-                        
-                        EditEntity(saveRoute, prefix, options.data.Id, editModalClass, modalWidth, refreshGrid);
-                    })
-                    .appendTo(container);
-            }
-        });
-    }
+    if (editModalClass || showDelete) {
 
-    if (showDelete) {
-        // add column for delete
         columns.push({
             width: '100px',
             alignment: 'center',
+            allowResizing: false,
             cellTemplate: function (container, options) {
-                $('<a/>')
-                    .addClass('deleteLink')
-                    .text('Delete')
-                    .click(function (e) {
-                        e.preventDefault();
-                        
-                        ConfirmModal('Are you sure you want to delete this item?', function () {
-                            DeleteEntity(saveRoute, options.data.Id, refreshGrid);
-                        }, null);
-                    })
-                    .appendTo(container);
+                if (editModalClass) {
+                    // Add button for edit
+
+                    $('<a/>')
+                        .addClass('actionbuttons')
+                        .addClass('editbutton')
+                        .attr('title', 'Edit')
+                        .click(function (e) {
+                            e.preventDefault();
+
+                            EditEntity(saveRoute, prefix, options.data.Id, editModalClass, modalWidth, refreshGrid);
+                        })
+                        .appendTo(container);
+                }
+
+                if (showDelete) {
+                    // Add button for delete
+
+                    $('<a/>')
+                        .addClass('actionbuttons')
+                        .addClass('deletebutton')
+                        .attr('title', 'Delete')
+                        .click(function (e) {
+                            e.preventDefault();
+
+                            ConfirmModal('Are you sure you want to delete this item?', function () {
+                                DeleteEntity(saveRoute, options.data.Id, refreshGrid);
+                            }, null);
+                        })
+                        .appendTo(container);
+                }
             }
         });
+
     }
 
     LoadGridData(container, gridClass, columns, getRoute, selected, newlink, showFilter, showGroup, onComplete);
@@ -272,6 +279,7 @@ function LoadGridData(container, grid, columns, getRoute, selected, newlink, sho
                 visible: showFilter,
                 showOperationChooser: false
             },
+            allowColumnResizing: true,
             selection: {
                 mode: 'single',
                 allowSelectAll: false
@@ -330,40 +338,45 @@ function LoadGridWithData(grid, container, columns, route, selected, editMethod,
 
     var datagrid = $('<div>').addClass(grid);
 
-    if (editMethod) {
-        columns.push({
-            width: '100px',
-            alignment: 'center',
-            cellTemplate: function(container, options) {
-                $('<a/>')
-                    .addClass('editthing')
-                    .text('Edit')
-                    .click(function(e) {
-                        e.preventDefault();
+    if (editMethod || deleteMethod) {
 
-                        editMethod(options.data.Id);
-                    })
-                    .appendTo(container);
-            }
-        });
-    }
-
-    if (deleteMethod) {
         columns.push({
             width: '100px',
             alignment: 'center',
             cellTemplate: function (container, options) {
-                $('<a/>')
-                    .addClass('editthing')
-                    .text('Delete')
-                    .click(function (e) {
-                        e.preventDefault();
 
-                        deleteMethod(options.data.Id);
-                    })
-                    .appendTo(container);
+                if (editMethod) {
+                    
+                    $('<a/>')
+                        .addClass('actionbuttons')
+                        .addClass('editbutton')
+                        .attr('title', 'Edit')
+                        .click(function (e) {
+                            e.preventDefault();
+
+                            editMethod(options.data.Id);
+                        })
+                        .appendTo(container);
+                }
+
+                if (deleteMethod) {
+
+                    $('<a/>')
+                        .addClass('actionbuttons')
+                        .addClass('deletebutton')
+                        .attr('title', 'Delete')
+                        .click(function (e) {
+                            e.preventDefault();
+
+                            deleteMethod(options.data.Id);
+                        })
+                        .appendTo(container);
+
+                }
+
             }
         });
+
     }
 
     var actualData = data;
@@ -397,10 +410,11 @@ function LoadGridWithData(grid, container, columns, route, selected, editMethod,
         filterRow: {
             visible: true,
             showOperationChooser: false
-            },
-            selection: {
-                mode: 'single',
-                allowSelectAll: false
+        },
+        allowColumnResizing: true,
+        selection: {
+            mode: 'single',
+            allowSelectAll: false
         },
         onRowClick: function (info) {
